+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/task/Task.java
@@ -17,4 +17,6 @@ public interface Task extends RootStorageListener {
     Collection<ResourceAmount> copyInternalStorageState();
 
     void step(RootStorage rootStorage, ExternalPatternInputSink externalPatternInputSink, StepBehavior stepBehavior);
+
+    void cancel();
 }