+++ refinedstorage-network/src/main/java/com/refinedmods/refinedstorage/api/network/impl/node/relay/RelayOutputPatternProvider.java
@@ -2,6 +2,7 @@ package com.refinedmods.refinedstorage.api.network.impl.node.relay;
 
 import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.autocrafting.task.Task;
+import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.core.Action;
 import com.refinedmods.refinedstorage.api.network.autocrafting.AutocraftingNetworkComponent;
 import com.refinedmods.refinedstorage.api.network.autocrafting.ParentContainer;
@@ -95,6 +96,11 @@ class RelayOutputPatternProvider implements PatternProvider, PatternListener {
 
     @Override
     public void addTask(final Task task) {
+        // TODO(feat): relay support
+    }
+
+    @Override
+    public void cancelTask(final TaskId taskId) {
         // TODO(feat): relay support
     }