+++ refinedstorage-network/src/test/java/com/refinedmods/refinedstorage/api/network/impl/autocrafting/AutocraftingNetworkComponentImplTest.java
@@ -4,13 +4,19 @@ import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewItem;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewType;
+import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
+import com.refinedmods.refinedstorage.api.autocrafting.task.TaskState;
 import com.refinedmods.refinedstorage.api.core.Action;
+import com.refinedmods.refinedstorage.api.network.Network;
+import com.refinedmods.refinedstorage.api.network.impl.NetworkImpl;
 import com.refinedmods.refinedstorage.api.network.impl.node.patternprovider.PatternProviderNetworkNode;
 import com.refinedmods.refinedstorage.api.network.node.container.NetworkNodeContainer;
+import com.refinedmods.refinedstorage.api.network.storage.StorageNetworkComponent;
+import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.StorageImpl;
 import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
-import com.refinedmods.refinedstorage.api.storage.root.RootStorageImpl;
+import com.refinedmods.refinedstorage.network.test.fixtures.NetworkTestFixtures;
 
 import java.util.Collections;
 import java.util.List;
@@ -27,12 +33,14 @@ import static com.refinedmods.refinedstorage.network.test.fixtures.ResourceFixtu
 import static org.assertj.core.api.Assertions.assertThat;
 
 class AutocraftingNetworkComponentImplTest {
+    private Network network;
     private RootStorage rootStorage;
     private AutocraftingNetworkComponentImpl sut;
 
     @BeforeEach
     void setUp() {
-        rootStorage = new RootStorageImpl();
+        network = new NetworkImpl(NetworkTestFixtures.NETWORK_COMPONENT_MAP_FACTORY);
+        rootStorage = network.getComponent(StorageNetworkComponent.class);
         sut = new AutocraftingNetworkComponentImpl(() -> rootStorage, Executors.newSingleThreadExecutor());
     }
 
@@ -122,11 +130,119 @@ class AutocraftingNetworkComponentImplTest {
         sut.onContainerAdded(container);
 
         // Act
-        final boolean success = sut.startTask(B, 1, Actor.EMPTY, false).join();
+        final Optional<TaskId> taskId = sut.startTask(B, 1, Actor.EMPTY, false).join();
 
         // Assert
-        assertThat(success).isTrue();
+        assertThat(taskId).isPresent();
         assertThat(provider.getTasks()).hasSize(1);
+    }
+
+    @Test
+    void shouldCancelTask() {
+        // Arrange
+        rootStorage.addSource(new StorageImpl());
+        rootStorage.insert(A, 10, Action.EXECUTE, Actor.EMPTY);
+
+        final PatternProviderNetworkNode provider1 = new PatternProviderNetworkNode(0, 5);
+        provider1.setActive(true);
+        provider1.setNetwork(network);
+        provider1.setPattern(1, pattern().ingredient(A, 3).output(B, 1).build());
+        sut.onContainerAdded(() -> provider1);
+
+        final PatternProviderNetworkNode provider2 = new PatternProviderNetworkNode(0, 5);
+        provider2.setActive(true);
+        provider2.setNetwork(network);
+        provider2.setPattern(1, pattern().ingredient(A, 3).output(C, 1).build());
+        sut.onContainerAdded(() -> provider2);
+
+        final Optional<TaskId> taskId1 = sut.startTask(B, 1, Actor.EMPTY, false).join();
+        final Optional<TaskId> taskId2 = sut.startTask(C, 1, Actor.EMPTY, false).join();
+
+        assertThat(taskId1).isPresent();
+        assertThat(taskId2).isPresent();
+        assertThat(provider1.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.READY);
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.READY);
+
+        // Act & assert
+        provider1.doWork();
+        provider2.doWork();
+        assertThat(provider1.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 4)
+        );
+
+        sut.cancel(taskId1.get());
+        assertThat(provider1.getTasks()).hasSize(1)
+            .allMatch(t -> t.getState() == TaskState.RETURNING_INTERNAL_STORAGE);
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 4)
+        );
+
+        provider1.doWork();
+        assertThat(provider1.getTasks()).isEmpty();
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 7)
+        );
+
+        sut.cancel(taskId1.get());
+    }
+
+    @Test
+    void shouldCancelAllTasks() {
+        // Arrange
+        rootStorage.addSource(new StorageImpl());
+        rootStorage.insert(A, 10, Action.EXECUTE, Actor.EMPTY);
+
+        final PatternProviderNetworkNode provider1 = new PatternProviderNetworkNode(0, 5);
+        provider1.setActive(true);
+        provider1.setNetwork(network);
+        provider1.setPattern(1, pattern().ingredient(A, 3).output(B, 1).build());
+        sut.onContainerAdded(() -> provider1);
+
+        final PatternProviderNetworkNode provider2 = new PatternProviderNetworkNode(0, 5);
+        provider2.setActive(true);
+        provider2.setNetwork(network);
+        provider2.setPattern(1, pattern().ingredient(A, 3).output(C, 1).build());
+        sut.onContainerAdded(() -> provider2);
+
+        final Optional<TaskId> taskId1 = sut.startTask(B, 1, Actor.EMPTY, false).join();
+        final Optional<TaskId> taskId2 = sut.startTask(C, 1, Actor.EMPTY, false).join();
+
+        assertThat(taskId1).isPresent();
+        assertThat(taskId2).isPresent();
+        assertThat(provider1.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.READY);
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.READY);
+
+        // Act & assert
+        provider1.doWork();
+        provider2.doWork();
+        assertThat(provider1.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(provider2.getTasks()).hasSize(1).allMatch(t -> t.getState() == TaskState.RUNNING);
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 4)
+        );
+
+        sut.cancelAll();
+        assertThat(provider1.getTasks()).hasSize(1)
+            .allMatch(t -> t.getState() == TaskState.RETURNING_INTERNAL_STORAGE);
+        assertThat(provider2.getTasks()).hasSize(1)
+            .allMatch(t -> t.getState() == TaskState.RETURNING_INTERNAL_STORAGE);
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 4)
+        );
+
+        provider1.doWork();
+        provider2.doWork();
+        assertThat(provider1.getTasks()).isEmpty();
+        assertThat(provider2.getTasks()).isEmpty();
+        assertThat(rootStorage.getAll()).usingRecursiveFieldByFieldElementComparator().containsExactly(
+            new ResourceAmount(A, 10)
+        );
+
+        sut.cancelAll();
     }
 
     @Test
@@ -138,10 +254,10 @@ class AutocraftingNetworkComponentImplTest {
         sut.onContainerAdded(container);
 
         // Act
-        final boolean success = sut.startTask(B, 2, Actor.EMPTY, false).join();
+        final Optional<TaskId> taskId = sut.startTask(B, 2, Actor.EMPTY, false).join();
 
         // Assert
-        assertThat(success).isFalse();
+        assertThat(taskId).isEmpty();
         assertThat(provider.getTasks()).isEmpty();
     }
 }