+++ CHANGELOG.md
@@ -18,6 +18,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Autocrafting now handles multiple patterns with the same output correctly by trying to use the pattern with the
     highest priority first. If there are missing resources, lower priority patterns are checked.
 -   The Autocrafter now faces the block you're clicking when placing it, like the other cable blocks (like the Exporter or Importer).
+-   You can no longer cancel autocrafting tasks if there is not enough space in storage to return the intermediate task storage.
 
 ### Fixed